reset examples according to `main`
