+++ examples/.interop/next-prisma-starter-websockets/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@examples/interop-next-starter-websockets",
+  "name": "@examples/legacy-trpc-next-starter-websockets",
   "version": "10.0.0-alpha.36",
   "private": true,
   "scripts": {
